bot/requirements.txt
flask
requests
redis
python-dotenv
upstash_redis
pytz
datetime